requirements.txt
Django==1.7.2
celery==3.1.17
djangorestframework==3.0.3
psycopg2==2.5.4
redis